+++ av1/decoder/obu.h
@@ -29,11 +29,11 @@ aom_codec_err_t aom_read_obu_header(uint8_t *buffer, size_t buffer_length,
                                     int is_annexb);
 
 aom_codec_err_t aom_read_obu_header_and_size(const uint8_t *data,
-                                             size_t bytes_available,
+                                             uint64_t bytes_available,
                                              int is_annexb,
                                              ObuHeader *obu_header,
-                                             size_t *const payload_size,
-                                             size_t *const bytes_read);
+                                             uint64_t *const payload_size,
+                                             uint64_t *const bytes_read);
 
 void aom_decode_frame_from_obus(struct AV1Decoder *pbi, const uint8_t *data,
                                 const uint8_t *data_end,
